+++ show.cpp
@@ -28,10 +28,10 @@ void showrole() {
 }
 
 void showpanel() {
-  Resize(&hp1, hpwidth * (role[0].hp / (double)initialhp), hpheight);
+  Resize(&hp1, (int)(hpwidth * (role[0].hp / (double)initialhp)), hpheight);
   drawAlpha(&gameimage.role_panel, 100, 35);
   drawAlpha(&hp1, 138, 54);
-  Resize(&hp2, hpwidth * (role[1].hp / (double)initialhp), hpheight);
+  Resize(&hp2, (int)(hpwidth * (role[1].hp / (double)initialhp)), hpheight);
   drawAlpha(&gameimage.role_panel, 890, 35);
   drawAlpha(&hp2, 928, 54);
